+++ lib/legion/extensions/cloudflare/accounts.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'legion/extensions/cloudflare/accounts/runners/accounts'
+require 'legion/extensions/cloudflare/accounts/runners/members'
+require 'legion/extensions/cloudflare/accounts/runners/roles'
+require 'legion/extensions/cloudflare/accounts/runners/tokens'
+require 'legion/extensions/cloudflare/accounts/client'
+
+module Legion
+  module Extensions
+    module Cloudflare
+      module Accounts
+      end
+    end
+  end
+end