+++ lib/legion/extensions/cloudflare/zones.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require 'legion/extensions/cloudflare/zones/runners/zones'
+require 'legion/extensions/cloudflare/zones/runners/settings'
+require 'legion/extensions/cloudflare/zones/client'
+
+module Legion
+  module Extensions
+    module Cloudflare
+      module Zones
+      end
+    end
+  end
+end